requirements.txt
langchain==0.3.27
langchain_core==0.3.72
langchain_google_genai==2.1.8
langchain_pinecone==0.2.11
nest_asyncio==1.5.6
nest_asyncio==1.6.0
pinecone==7.3.0
pypdf==5.9.0
python-dotenv==1.1.1
streamlit==1.47.1
